6335: Moved updateUserRecords, changed error to available on adminGetUser.
